// apps/admin-service/src/modules/users/entities/enums/user-role.enum.ts
export enum UserRole {
  SUPER_ADMIN = 'super_admin',
  CTO = 'cto',
  GROWTH_FINANCE = 'growth_finance',
  CUSTOMER_SUPPORT = 'customer_support',
  CONTENT_MANAGER = 'content_manager',
  CUSTOMER_MANAGER = 'customer_manager',
  FINANCIAL_ADMIN = 'financial_admin',
  COMPANY_ADMIN = 'company_admin',
  COMPANY_USER = 'company_user',
  // Add any other roles that might be relevant from other modules or general use
}